requirements.txt
flask==3.0.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
pymupdf==1.23.0
python-docx==0.8.11
gunicorn==21.2.0